LAB_1/LAB_1/requirements.txt
fastapi
uvicorn[standard]
httpie
requests
python-multipart 
pillow